requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP requests and web scraping
requests==2.31.0
httpx==0.25.2
cloudscraper==1.2.71
selenium==4.15.2
undetected-chromedriver==3.5.4

# Database
sqlalchemy==2.0.23
alembic==1.12.1
# sqlite3 is built-in to Python

# Data processing
pandas==2.1.3
numpy==1.25.2
beautifulsoup4==4.12.2
lxml==4.9.3

# Async support
# asyncio is built-in to Python
aiofiles==23.2.1
asyncpg==0.29.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
schedule==1.2.0
python-multipart==0.0.6

# Logging and monitoring
loguru==0.7.2

# Notifications
python-telegram-bot==20.7
# smtplib is built-in to Python

# Deployment
gunicorn==21.2.0
